Documents/practical_simprosys/ecommerce/requirements.txt
Django=5.0.1
djangorestframework=3.14.0
djangorestframework-simplejwt=5.3.1
celery=5.3.6
Faker=22.6.0
django-celery-results=2.5.1
django-redis=5.4.0
cryptography=42.0.1